% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotateGenomicCoordinates.R
\name{annotate_genomic_position}
\alias{annotate_genomic_position}
\title{Convenience function to annotate a single position}
\usage{
annotate_genomic_position(
  chr,
  pos,
  genome,
  gtffile,
  tss_upstream = 2000,
  tss_downstream = 200,
  cache_gtf = TRUE,
  include_exon_info = FALSE
)
}
\description{
Convenience function to annotate a single position
}
\keyword{internal}
