% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{generate_offsets}
\alias{generate_offsets}
\title{Generate Cumulative Genomic Offsets}
\usage{
generate_offsets(coverage_data, variant_calls)
}
\arguments{
\item{coverage_data}{A data frame with columns \code{seqnames}, \code{start}, and \code{end}
representing coverage intervals.}

\item{variant_calls}{A data frame with columns \code{seqnames}, \code{start}, and \code{end}
representing variant calls.}
}
\value{
A list with two elements:
  \describe{
    \item{\code{coverage_data}}{The input coverage data augmented with a column
      \code{cumulative_genomic_position}.}
    \item{\code{variant_calls}}{The input variant calls augmented with columns
      \code{cumulative_start} and \code{cumulative_end}.}
  }
}
\description{
Computes cumulative genomic positions for coverage data and variant calls by calculating
chromosome offsets.
}
\details{
The function filters for valid chromosomes (chr1 - chr22, X, Y), calculates the maximum
  \code{end} of each chromosome as a proxy for its length, and uses the cumulative sum of chromosome
  lengths to determine an offset for each chromosome. It then adds this offset to the \code{start} of
  each interval to obtain a cumulative genomic position.
}
\keyword{internal}
