% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pacbiowdlR-package.R
\docType{package}
\name{pacbiowdlR-package}
\alias{pacbiowdlR}
\alias{pacbiowdlR-package}
\title{pacbiowdlR: Plotting DNA Copy Number Alterations from PacBio WDL Data}
\description{
The **pacbiowdlR** package provides functions for visualizing copy number alterations (CNAs)
from PacBio WDL pipelines. Its functions leverage **ggplot2** to create customizable plots.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://scfurl.github.io/pacbiowdlR/}
}

}
\author{
\strong{Maintainer}: Scott Furlan \email{scottfurlan@gmail.com} (\href{https://orcid.org/0000-0003-3007-9840}{ORCID})

}
