% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotateGenomicCoordinates.R
\name{preload_gtf}
\alias{preload_gtf}
\title{Preload GTF data with proper symbol handling and canonical transcript tracking}
\usage{
preload_gtf(gtffile)
}
\description{
Preload GTF data with proper symbol handling and canonical transcript tracking
}
\keyword{internal}
