% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fusionplot2.R
\name{run_fusion_examples}
\alias{run_fusion_examples}
\title{Run an interactive fusion visualization demo}
\usage{
run_fusion_examples()
}
\description{
Demonstrates the fusion visualization with different examples
}
