% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate_genomic_coordinates.R
\name{load_cached_gtf}
\alias{load_cached_gtf}
\title{Load cached GTF data or rebuild if needed}
\usage{
load_cached_gtf(gtffile, force = FALSE)
}
\description{
Load cached GTF data or rebuild if needed
}
\keyword{internal}
