% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pacbiowdlR-package.R
\name{pacbiowdlR-package}
\alias{pacbiowdlR-package}
\title{pacbiowdlR: Plotting DNA Copy Number Alterations from PacBio WDL Data}
\description{
The **pacbiowdlR** package provides functions for visualizing copy number alterations (CNAs)
from PacBio WDL pipelines. Its functions leverage **ggplot2** to create customizable plots.
}
\details{
_PACKAGE package
}
